fix(test): improve CtGenerationTest and CtScannerTest (#1736)

+++ src/test/java/spoon/generating/CloneVisitorGenerator.java
@@ -100,7 +100,11 @@ public class CloneVisitorGenerator extends AbstractManualProcessor {
 
 				// Changes body of the cloned method.
 				for (int i = 1; i < clone.getBody().getStatements().size() - 1; i++) {
-					final CtInvocation targetInvocation = (CtInvocation) ((CtInvocation) clone.getBody().getStatement(i)).getArguments().get(1);
+					List<CtExpression> invArgs = ((CtInvocation) clone.getBody().getStatement(i)).getArguments();
+					if (invArgs.size() <= 1) {
+						throw new RuntimeException("You forget the role argument in line "+i+" of method "+element.getSimpleName()+" from "+element.getDeclaringType().getQualifiedName());
+					}
+					final CtInvocation targetInvocation = (CtInvocation) invArgs.get(1);
 					if ("getValue".equals(targetInvocation.getExecutable().getSimpleName()) && "CtLiteral".equals(targetInvocation.getExecutable().getDeclaringType().getSimpleName())) {
 						clone.getBody().getStatement(i--).delete();
 						continue;
